Remove `metal/objc2.rs`, inline its content into `metal.rs`, and adjust imports in `unity-native-plugin-sample` to simplify structure and avoid redundant module.
